red_lab_03_sniffing/app/requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.7
bcrypt==4.0.1
gunicorn==20.1.0
